Renamed some fields and refactored repositories
