--- src/main/java/com/refactoringflow/refactoringflowbackend/Exception/FileUploadSizeException.java
@@ -1,18 +1,0 @@
-package com.refactoringflow.refactoringflowbackend.Exception;
-
-import com.refactoringflow.refactoringflowbackend.ResponseMessage;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ControllerAdvice;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.multipart.MaxUploadSizeExceededException;
-import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
-
-@ControllerAdvice
-public class FileUploadSizeException extends ResponseEntityExceptionHandler {
-    @ExceptionHandler(MaxUploadSizeExceededException.class)
-    public ResponseEntity<ResponseMessage> handleMaxSizeException(MaxUploadSizeExceededException ex){
-        return ResponseEntity.status(HttpStatus.EXPECTATION_FAILED).body(new ResponseMessage("File size is to big!"));
-    }
-
-}
