--- src/main/java/com/refactoringflow/refactoringflowbackend/Exception/ResourceNotFoundException.java
@@ -1,14 +1,0 @@
-package com.refactoringflow.refactoringflowbackend.Exception;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-@ResponseStatus(value = HttpStatus.NOT_FOUND)
-public class ResourceNotFoundException extends RuntimeException{
-
-    private static final long serialVersionUID = 1L;
-
-    public ResourceNotFoundException(String message) {
-        super(message);
-    }
-}
