--- src/main/java/com/refactoringflow/refactoringflowbackend/Model/Course.java
@@ -1,24 +1,0 @@
-package com.refactoringflow.refactoringflowbackend.Model;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-import javax.persistence.*;
-
-@Getter
-@Setter
-@AllArgsConstructor
-@NoArgsConstructor
-@Entity
-@Table(name = "Course")
-public class Course {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long ID;
-
-    @Column(name = "Title")
-    private String Title;
-}
