--- src/main/java/com/refactoringflow/refactoringflowbackend/Model/LearningOutcomes.java
@@ -1,25 +1,0 @@
-package com.refactoringflow.refactoringflowbackend.Model;
-
-import lombok.*;
-
-import javax.persistence.*;
-
-@Getter
-@Setter
-@AllArgsConstructor
-@NoArgsConstructor
-@Entity
-@Table(name = "learning_outcomes")
-public class LearningOutcomes {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long Id;
-    @OneToOne(cascade = CascadeType.MERGE)
-    @JoinColumn(name = "student_id", referencedColumnName = "id")
-    private Student student;
-    @Column(name = "code_quality")
-    private long codeQuality;
-    @Column(name = "refactoring")
-    private long refactoring;
-}
