--- src/main/java/com/refactoringflow/refactoringflowbackend/Repository/SegmentRepository.java
@@ -1,15 +1,0 @@
-package com.refactoringflow.refactoringflowbackend.Repository;
-
-
-import com.refactoringflow.refactoringflowbackend.Model.Course;
-import com.refactoringflow.refactoringflowbackend.Model.Segment;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface SegmentRepository extends JpaRepository<Segment, Long> {
-    Optional<Segment> findByCourseId(long CourseId);
-    Segment findByNrAndCourseId(Long Nr, Course course);
-}
