--- src/main/java/com/refactoringflow/refactoringflowbackend/Repository/StudentRepository.java
@@ -1,11 +1,0 @@
-package com.refactoringflow.refactoringflowbackend.Repository;
-
-
-import com.refactoringflow.refactoringflowbackend.Model.Student;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface StudentRepository extends JpaRepository<Student, Long> {
-    Student findStudentByEmailAndPassword(String email, String password);
-}
