+++ src/main/java/com/refactoringflow/refactoringflowbackend/controller/StudentController.java
@@ -1,0 +1,42 @@
+package com.refactoringflow.refactoringflowbackend.controller;
+
+import com.refactoringflow.refactoringflowbackend.model.Student;
+import com.refactoringflow.refactoringflowbackend.service.StudentService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
+import java.util.Optional;
+
+@RestController
+@CrossOrigin
+@RequestMapping("api/v1/student")
+public class StudentController {
+
+    StudentService studentService;
+
+    @Autowired
+    public StudentController(StudentService studentService) {
+        this.studentService = studentService;
+    }
+
+    @GetMapping("/")
+    public List<Student> findAllStudents(){
+        return studentService.findAll();
+    }
+
+    @GetMapping("/{id}")
+    public Optional<Student> findById(@PathVariable Long id){
+        return studentService.findById(id);
+    }
+
+    @PostMapping("/update")
+    public Student updateStudent(@RequestBody Student student){
+        return studentService.updateStudent(student);
+    }
+
+    @GetMapping("/login/{email}/{password}")
+    public Student login(@PathVariable String email, @PathVariable String password){
+        return studentService.findStudentByEmailAndPassword(email, password);
+    }
+}
