+++ src/main/java/com/refactoringflow/refactoringflowbackend/model/Course.java
@@ -1,0 +1,24 @@
+package com.refactoringflow.refactoringflowbackend.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+import javax.persistence.*;
+
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+@Entity
+@Table(name = "Course")
+public class Course {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private long ID;
+
+    @Column(name = "Title")
+    private String Title;
+}
