+++ src/main/java/com/refactoringflow/refactoringflowbackend/repository/CodeFileRepository.java
@@ -1,0 +1,10 @@
+package com.refactoringflow.refactoringflowbackend.repository;
+
+import com.refactoringflow.refactoringflowbackend.model.CodeFile;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface CodeFileRepository extends JpaRepository<CodeFile, Long> {
+
+}
