+++ src/main/java/com/refactoringflow/refactoringflowbackend/repository/CourseRepository.java
@@ -1,0 +1,7 @@
+package com.refactoringflow.refactoringflowbackend.repository;
+
+import com.refactoringflow.refactoringflowbackend.model.Course;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface CourseRepository extends JpaRepository<Course, Long> {
+}
