+++ src/main/java/com/refactoringflow/refactoringflowbackend/repository/StudentRepository.java
@@ -1,0 +1,11 @@
+package com.refactoringflow.refactoringflowbackend.repository;
+
+
+import com.refactoringflow.refactoringflowbackend.model.Student;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface StudentRepository extends JpaRepository<Student, Long> {
+    Student findStudentByEmailAndPassword(String email, String password);
+}
