+++ src/main/java/com/refactoringflow/refactoringflowbackend/repository/TeacherRepository.java
@@ -1,0 +1,10 @@
+package com.refactoringflow.refactoringflowbackend.repository;
+
+import com.refactoringflow.refactoringflowbackend.model.Teacher;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface TeacherRepository extends JpaRepository<Teacher, Long> {
+    Teacher findTeacherByEmailAndPassword(String email, String password);
+}
